requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.11
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
flake8==5.0.4
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.10.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
htmlmin==0.1.12
hydra-core==1.2.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.2
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
missingno==0.5.1
mistune==2.0.4
multidict==6.0.2
multimethod==1.9
mypy==0.982
mypy-extensions==0.4.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
notebook==6.5.1
boto3==1.17.49
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
pandas-profiling==3.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
patsy==0.5.3
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
pyflakes==2.5.0
Pygments==2.13.0
pyyaml==3.11
pylint==2.15.4
dataclasses==0.7
python-dotenv>=0.5.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.7.7
pytz==2022.5
PyWavelets==1.4.1
marshmallow-dataclass==8.3.0
PyYAML==6.0
mlflow==1.15.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
statsmodels==0.13.2
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.5
torch==1.12.1
torchmetrics==0.10.0
torchvision==0.13.1
tornado==6.2
tox==3.26.0
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
visions==0.7.5
watermark==2.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
